requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
httpx
pillow
python-dotenv
python-slugify
google-genai
